+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixes
 
 - Resolve relative `SOURCEMAP_FILE` paths against the project root in the Xcode build script ([#5730](https://github.com/getsentry/sentry-react-native/pull/5730))
+- Fixes the issue with unit mismatch in `adjustTransactionDuration` ([#5740](https://github.com/getsentry/sentry-react-native/pull/5740))
 - Handle `inactive` state for spans ([#5742](https://github.com/getsentry/sentry-react-native/pull/5742))
 
 ### Dependencies